Setup a backend dir with server.js
